# api/swagger.yml
openapi: 3.0.3
info:
  title: 'Webapp swagger - OpenAPI 3.0'
  description: "This is a Webapp swagger"
  version: 0.0.1 (v3)
servers:
  - url: https://staging.jacekplaymaker.usermd.net/api/v3/
  - url: https://dev.jacekplaymaker.usermd.net/api/v3/
  - url: http://localhost:{port}/api/v3/
    variables:
      port:
        default: '8000'
paths:
...
  /inquiries/my/contacts/:
    get:
      operationId: get_my_contacts
      description: Get my contacts from inquiry requests
      summary: Get my contacts from inquiry requests
      tags:
        - inquiries
      security:
        - jwtAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/X-Language'
      responses:
        '200':
          description: "List of accepted inquiry requests (as sender and recipient)"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InquiryRequest'
        '401':
          description: "Unauthorized"
...
  /inquiries/{iq_id}/accept/:
    post:
      operationId: accept_inquiry_request
      description: Accept inquiry request
      summary: Accept inquiry request
      parameters:
        - in: path
          name: iq_id
          schema:
            type: integer
            description: InquiryRequest ID
          required: true
        - $ref: '#/components/parameters/X-Language'
      tags:
        - inquiries
      security:
        - jwtAuth: [ ]
      responses:
        '200':
          description: "Inquiry request accepted successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InquiryRequest'
        '401':
          description: "Unauthorized"
        '404':
          description: "Inquiry request with given ID doesn't exist"
        '400':
          description: "You cannot accept your own inquiry request"
  /inquiries/{iq_id}/reject/:
    post:
      operationId: reject_inquiry_request
      description: Reject inquiry request
      summary: Reject inquiry request
      parameters:
        - in: path
          name: iq_id
          schema:
            type: integer
            description: InquiryRequest ID
          required: true
        - $ref: '#/components/parameters/X-Language'
      tags:
        - inquiries
      security:
        - jwtAuth: [ ]
      responses:
        '200':
          description: "Inquiry request rejected successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InquiryRequest'
        '401':
          description: "Unauthorized"
        '404':
          description: "Inquiry request with given ID doesn't exist"
        '400':
          description: "You cannot reject your own inquiry request"
...
  /followers/profile/{profile_uuid}/follow/:
    post:
      operationId: follow_profile
      description: "Follow a profile identified by its UUID."
      summary: "Follow a profile"
      tags:
        - followers
      security:
        - jwtAuth: [ ]
      parameters:
        - name: profile_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: "The UUID of the profile to follow"
        - $ref: '#/components/parameters/X-Language'
      responses:
        '201':
          description: "Profile followed successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Profile followed successfully"
        '401':
          description: "Unauthorized - Authentication credentials were not provided or are invalid"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Authentication credentials were not provided or are invalid."
        '400':
          description: "Bad Request - You cannot follow your own profile."
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "You cannot follow your own profile."
        '403':
          description: "Forbidden - You are already following this profile."
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "You are already following this profile."
        '404':
          description: "Not Found - The requested profile was not found."
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Profile with given UUID does not exist."
  /followers/profile/{profile_uuid}/unfollow/:
    delete:
      operationId: unfollow_profile
      description: "Unfollow a profile identified by its UUID."
      summary: "Unfollow a profile"
      tags:
        - followers
      security:
        - jwtAuth: [ ]
      parameters:
        - name: profile_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: "The UUID of the profile to unfollow"
        - $ref: '#/components/parameters/X-Language'
      responses:
        '204':
          description: "Profile unfollowed successfully"
        '401':
          description: "Unauthorized - User not authenticated."
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Authentication credentials were not provided or are invalid."
        '403':
          description: "Forbidden - You do not have permission to perform this action."
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "You do not have permission to perform this action."
        '404':
          description: "Not Found - The specified profile or follow relationship does not exist."
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "The specified profile or follow relationship does not exist."
...
components:
  schemas:
...
    BaseUserData:
      type: object
      properties:
        id:
          type: integer
        first_name:
          type: string
        last_name:
          type: string
        last_login:
          type: string
        picture:
          type: string
        role:
          type: string
      required:
        - id
...
    Gender:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required:
        - id
        - name
...
    TeamHistoryProfileView:
      type: object
      required:
        - team_name
        - league_name
        - league_id
        - team_contributor_id
        - picture_url
        - season
        - country
      properties:
        team_name:
          type: string
        league_name:
          type: string
        league_id:
          type: integer
        team_contributor_id:
          type: integer
        picture_url:
          type: string
          nullable: true
          format: uri
        country:
          type: string
        season:
          type: string
...
    Position:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        shortcut:
          type: string
      required:
        - id
        - name
        - shortcut
    PlayerPosition:
      type: object
      properties:
        player_position:
          $ref: '#/components/schemas/Position'
        is_main:
          type: boolean
      required:
        - player_position
        - is_main
...
    PlayerProfileMetrics:
      type: object
      properties:
        pm_score:
          type: number
          readOnly: true
        season_score:
          type: number
          readOnly: true
        pm_score_state:
          type: string
          readOnly: true
        pm_score_history:
          type: object
        pm_score_change:
          type: number

      required:
        - pm_score
        - season_score
        - pm_score_state
...
    ProfileEnumChoices:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      description: 'Object represents enums, common for all models'
      required:
        - id
        - name
...
    InquirySpecifiedUserBaseData:
      allOf:
        - $ref: '#/components/schemas/BaseUserData'
        - type: object
          properties:
            uuid:
              type: string
              format: uuid
            slug:
              type: string
              readOnly: true
            custom_role:
              type: string
            specific_role:
              $ref: '#/components/schemas/ProfileEnumChoices'
            age:
              type: integer
            player_position:
              $ref: '#/components/schemas/PlayerPosition'
            team_history_object:
              $ref: '#/components/schemas/TeamHistoryProfileView'
            gender:
              $ref: '#/components/schemas/Gender'
            contact:
              $ref: '#/components/schemas/ContactSchema'
            playermetrics:
              $ref: '#/components/schemas/PlayerProfileMetrics'
    InquiryRequest:
      type: object
      properties:
        id:
          type: integer
          description: Inquiry id
          readOnly: true
        created_at:
          type: string
          readOnly: true
        updated_at:
          type: string
          readOnly: true
        status:
          type: string
          readOnly: true
        sender_object:
          $ref: '#/components/schemas/InquirySpecifiedUserBaseData'
        recipient_object:
          $ref: '#/components/schemas/InquirySpecifiedUserBaseData'
        sender:
          type: integer
          description: Sender ID
        recipient:
          type: integer
          description: Recipient ID
        is_read_by_sender:
          type: boolean
          description: Indicates whether the sender of the inquiry has viewed the recipient's response
        is_read_by_recipient:
          type: boolean
          description: Specifies whether the inquiry has been read by the recipient
      required:
        - id
        - created_at
        - updated_at
        - status
        - sender
        - recipient
        - is_read_by_sender
        - is_read_by_recipient
...
    ContactSchema:
      type: object
      properties:
        phone_number:
          type: object
          properties:
            number:
              type: string
            dial_code:
              type: string
        email:
          type: string
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT